# perf: r1cs solver 40% faster by avoiding redudnant check