… go.sum
… internal/backend/bls12-377/cs/r1cs_test.go
… internal/backend/bls12-381/cs/r1cs_test.go
… internal/backend/bls24-315/cs/r1cs_test.go
… internal/backend/bls24-315/cs/solution.go
… internal/backend/bn254/cs/r1cs_test.go
… internal/backend/bn254/cs/solution.go
… internal/backend/bw6-633/cs/r1cs.go
… internal/backend/bw6-633/cs/r1cs_test.go
… internal/backend/bw6-633/cs/solution.go
… internal/backend/bw6-761/cs/r1cs.go
… internal/backend/bw6-761/cs/r1cs_test.go
… internal/backend/bw6-761/cs/solution.go
… internal/generator/backend/template/representations/r1cs.go.tmpl
… internal/generator/backend/template/representations/solution.go.tmpl
… internal/generator/backend/template/representations/tests/r1cs.go.tmpl