- test: added r1cs solve benchmark
- perf: r1cs solver 40% faster by avoiding redudnant check
- fix: r1cs solver was incorrectly doing a multiplication by a coeff instead of a div. It didn't show since coeff value at this stage was -1 or 1 (== same result for div). 

```
name      old time/op    new time/op    delta
Solve-10    61.2ms ± 2%    37.3ms ± 2%  -39.07%  (p=0.333 n=2+2)
```